Split WindowPartitionKeys grab-bag header
WindowPartitionKeys.h bundled three unrelated types in the detail namespace
even though they were used across translation units. Separate them by
responsibility:

- RowColumnsSnapshot (renamed from WindowPartitionKeyRowSnapshot) is the only
  genuinely shared type. It copies a subset of a row's columns so the row can
  be compared after its source vector is gone, which is not window-specific.
  It moves to RowColumnsSnapshot.{h,cpp} in the exec namespace.
- The reference into a retained input vector is used only by
  VectorWindowPartition, so it moves to an anonymous namespace in the .cpp
  together with rowAt and rowsEqual.
- The key-channel deduplication becomes a small free function inlined into
  each of the two callers instead of a class with static methods.

No behavior change.
